simple-rag/simple-agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
yfinance>=0.2.18
requests>=2.31.0
pytest>=7.4.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
plotly>=5.17.0
streamlit>=1.29.0
crewai>=0.5.0